fixed compilation warnings about signed/unsigned comparisons fixed compilation warnings about const string literals some cosmetic fixes

+++ arabic.c
@@ -196,7 +196,7 @@ static const unsigned short transparent[] = {
 static int is_transparent(int ch)
 {
     int i;
-    for(i=0;i<sizeof(transparent) / sizeof(transparent[0]);i++) {
+    for (i = 0; i < (int)(sizeof(transparent)/sizeof(transparent[0])); i++) {
         if (transparent[i] == ch)
             return 1;
     }
